Make more modest performance claims
While the previous claim was true on the laptop I was developing this
on, I did some more benchmarking on other machines since then, and on
some the overhead when computing hashes in isolation is quite a bit
larger.

Luckily this doesn't change the overall picture when using ZwoHash in a
hash table.